refactor reporting link helpers
